# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BillForge API",
    "description": "BillForge - Modern Invoice Processing Platform API\n\n## Overview\n\nBillForge is a comprehensive accounts payable automation platform that helps businesses:\n- Capture and digitize invoices with OCR\n- Route invoices through customizable approval workflows\n- Manage vendor relationships\n- Generate financial reports and analytics\n\n## Authentication\n\nAll API endpoints (except /health) require authentication using JWT Bearer tokens.\n\n1. Obtain tokens via POST /api/v1/auth/login\n2. Include the access token in the Authorization header: `Authorization: Bearer <token>`\n3. Refresh tokens before expiry using POST /api/v1/auth/refresh\n\n## Multi-Tenancy\n\nBillForge is a multi-tenant system. Include tenant_id in authentication requests.\n",
    "contact": {
      "name": "BillForge Support",
      "email": "support@billforge.app"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API Server"
    }
  ],
  "paths": {
...
    "/api/v1/analytics/benchmark": {
      "get": {
        "tags": [
          "Benchmarking"
        ],
        "summary": "`GET /api/v1/analytics/benchmark`",
        "description": "Returns the tenant's six AP KPIs alongside anonymized cohort percentiles.\nIf `benchmark_opt_in = false`, returns `{ opted_in: false }` so the UI\ncan render the opt-in CTA.",
        "operationId": "get_benchmark",
        "responses": {
          "200": {
            "description": "Benchmark KPIs and cohort percentiles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/analytics/benchmark/opt-in": {
      "post": {
        "tags": [
          "Benchmarking"
        ],
        "summary": "`POST /api/v1/analytics/benchmark/opt-in`",
        "description": "Toggles the benchmark_opt_in flag and stores the cohort descriptor.",
        "operationId": "benchmark_opt_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkOptInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Opt-in accepted; returns updated benchmark snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/billing/usage": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "GET /billing/usage - return per-tenant invoice and vendor counts for the current billing period",
        "operationId": "get_usage",
        "responses": {
          "200": {
            "description": "Current billing period usage"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/routing/invoices/{invoice_id}/route": {
      "post": {
        "tags": [
          "Routing"
        ],
        "summary": "Get a routing decision for an invoice",
        "operationId": "route_invoice",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "description": "Invoice ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Routing decision"
          },
          "404": {
            "description": "Invoice not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BenchmarkKpis": {
        "type": "object",
        "description": "Six AP KPIs for a single tenant or a cohort percentile bucket.",
        "required": [
          "dpo_days",
          "avg_approval_cycle_hours",
          "ocr_straight_through_rate",
          "exception_rate",
          "discount_capture_rate",
          "cost_per_invoice"
        ],
        "properties": {
          "avg_approval_cycle_hours": {
            "type": "number",
            "format": "double"
          },
          "cost_per_invoice": {
            "type": "number",
            "format": "double"
          },
          "discount_capture_rate": {
            "type": "number",
            "format": "double"
          },
          "dpo_days": {
            "type": "number",
            "format": "double"
          },
          "exception_rate": {
            "type": "number",
            "format": "double"
          },
          "ocr_straight_through_rate": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "BenchmarkOptInRequest": {
        "type": "object",
        "description": "Request body for `POST /api/analytics/benchmark/opt-in`.",
        "required": [
          "industry",
          "headcount_band",
          "volume_band"
        ],
        "properties": {
          "headcount_band": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          },
          "volume_band": {
            "type": "string"
          }
        }
      },
      "BenchmarkResponse": {
        "type": "object",
        "description": "Full benchmark response returned by `GET /api/analytics/benchmark`.",
        "required": [
          "opted_in"
        ],
        "properties": {
          "cohort": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CohortDescriptor"
              }
            ]
          },
          "cohort_kpis": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CohortPercentiles"
              }
            ]
          },
          "cohort_size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "opted_in": {
            "type": "boolean"
          },
          "tenant_kpis": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BenchmarkKpis"
              }
            ]
          }
        }
      },
...
      "CohortDescriptor": {
        "type": "object",
        "description": "Cohort descriptor (the three dimensions that define a peer group).",
        "required": [
          "industry",
          "headcount_band",
          "volume_band"
        ],
        "properties": {
          "headcount_band": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          },
          "volume_band": {
            "type": "string"
          }
        }
      },
      "CohortPercentiles": {
        "type": "object",
        "description": "Percentile bands for cohort KPIs.",
        "required": [
          "p25",
          "p50",
          "p75"
        ],
        "properties": {
          "p25": {
            "$ref": "#/components/schemas/BenchmarkKpis"
          },
          "p50": {
            "$ref": "#/components/schemas/BenchmarkKpis"
          },
          "p75": {
            "$ref": "#/components/schemas/BenchmarkKpis"
          }
        }
      },
...
    }
  },
...
}